Refactor stack to use constructs
