+++ .eslintrc.json
@@ -37,5 +37,6 @@
   "env": {
     "browser": true,
     "es2021": true
-  }
+  },
+  "ignorePatterns": ["esbuild.js"]
 }
