+++ .gitignore
@@ -2,9 +2,11 @@
 !tailwind.config.js
 !postcss.config.js
 !jest.config.js
+!esbuild.js
 *.d.ts
-dist/
 node_modules/
+/build
+/dist
 __pycache__/
 .ruff_cache/
 .husky/
@@ -31,7 +33,7 @@ coverage
 
 # Model Cache Folders
 /models
+
 # Deployment configuration file
 config.yaml
 config-custom.yaml
-/.coverage
