+++ lib/iam/iamConstruct.ts
@@ -16,22 +16,21 @@
 import { Stack, StackProps } from 'aws-cdk-lib';
 import { IManagedPolicy, IRole, ManagedPolicy, Role, ServicePrincipal } from 'aws-cdk-lib/aws-iam';
 import { StringParameter } from 'aws-cdk-lib/aws-ssm';
-import { NagSuppressions } from 'cdk-nag';
 import { Construct } from 'constructs';
 
-import { createCdkId, getIamPolicyStatements } from './core/utils';
-import { BaseProps, Config } from './schema';
-import { getRoleId, ROLE, Roles } from './core/iam/roles';
+import { createCdkId, getIamPolicyStatements } from '../core/utils';
+import { BaseProps, Config } from '../schema';
+import { getRoleId, ROLE, Roles } from '../core/iam/roles';
 
 /**
  * Properties for the LisaServeIAMStack Construct.
  */
-type LisaIAMStackProps = {} & BaseProps & StackProps;
+export type LisaIAMStackProps = BaseProps & StackProps;
 
 /**
  * Properties for the ECS Role definitions
  */
-type ECSRole = {
+export type ECSRole = {
     id: string;
     type: ECSTaskType;
 };
@@ -39,30 +38,26 @@ type ECSRole = {
 /**
  * ECS Task types
  */
-enum ECSTaskType {
+export enum ECSTaskType {
     API = 'API',
 }
 
 /**
- * LisaServe IAM stack.
+ * LisaServe IAM.
  */
-export class LisaServeIAMStack extends Stack {
+export class LisaServeIAMSConstruct extends Construct {
+
+    private readonly scope: Stack;
 
     /**
-     * @param {Construct} scope - The parent or owner of the construct.
+     * @param {Stack} scope - The parent or owner of the construct.
      * @param {string} id - The unique identifier for the construct within its scope.
      * @param {LisaIAMStackProps} props - Properties for the Stack.
      */
-    constructor (scope: Construct, id: string, props: LisaIAMStackProps) {
-        super(scope, id, props);
+    constructor (scope: Stack, id: string, props: LisaIAMStackProps) {
+        super(scope, id);
+        this.scope = scope;
         const { config } = props;
-        // Add suppression for IAM4 (use of managed policy)
-        NagSuppressions.addStackSuppressions(this, [
-            {
-                id: 'AwsSolutions-IAM4',
-                reason: 'Allow use of AmazonSSMManagedInstanceCore policy to allow EC2 to enable SSM core functionality.',
-            },
-        ]);
 
         /*
         * Create role for Lambda execution if deploying RAG
@@ -91,10 +86,10 @@ export class LisaServeIAMStack extends Stack {
             const taskRoleName = createCdkId([config.deploymentName, role.id, ROLE]);
             const taskRole = config.roles ?
                 // @ts-expect-error - dynamic key lookup of object
-                Role.fromRoleName(this, taskRoleId, config.roles[taskRoleOverride]) :
+                Role.fromRoleName(scope, taskRoleId, config.roles[taskRoleOverride]) :
                 this.createEcsTaskRole(role, taskRoleId, taskRoleName, taskPolicy);
 
-            new StringParameter(this, createCdkId([config.deploymentName, role.id, 'SP']), {
+            new StringParameter(scope, createCdkId([config.deploymentName, role.id, 'SP']), {
                 parameterName: `${config.deploymentPrefix}/roles/${role.id}`,
                 stringValue: taskRole.roleArn,
                 description: `Role ARN for LISA ${role.type} ${role.id} ECS Task`,
@@ -103,9 +98,9 @@ export class LisaServeIAMStack extends Stack {
             if (config.roles) {
                 const executionRoleOverride = getRoleId(`ECS_${role.id}_${role.type}_EX_ROLE`.toUpperCase());
                 // @ts-expect-error - dynamic key lookup of object
-                const executionRole = Role.fromRoleName(this, createCdkId([role.id, 'ExRole']), config.roles[executionRoleOverride]);
+                const executionRole = Role.fromRoleName(scope, createCdkId([role.id, 'ExRole']), config.roles[executionRoleOverride]);
 
-                new StringParameter(this, createCdkId([config.deploymentName, role.id, 'EX', 'SP']), {
+                new StringParameter(scope, createCdkId([config.deploymentName, role.id, 'EX', 'SP']), {
                     parameterName: `${config.deploymentPrefix}/roles/${role.id}EX`,
                     stringValue: executionRole.roleArn,
                     description: `Role ARN for LISA ${role.type} ${role.id} ECS Execution`,
@@ -117,12 +112,12 @@ export class LisaServeIAMStack extends Stack {
     private createTaskPolicy (deploymentName: string, deploymentPrefix?: string): IManagedPolicy {
         const statements = getIamPolicyStatements('ecs');
         const taskPolicyId = createCdkId([deploymentName, 'ECSPolicy']);
-        const taskPolicy = new ManagedPolicy(this, taskPolicyId, {
+        const taskPolicy = new ManagedPolicy(this.scope, taskPolicyId, {
             managedPolicyName: createCdkId([deploymentName, 'ECSPolicy']),
             statements,
         });
 
-        new StringParameter(this, createCdkId(['ECSPolicy', 'SP']), {
+        new StringParameter(this.scope, createCdkId(['ECSPolicy', 'SP']), {
             parameterName: `${deploymentPrefix}/policies/${taskPolicyId}`,
             stringValue: taskPolicy.managedPolicyArn,
             description: `Managed Policy ARN for LISA ${taskPolicyId}`,
@@ -134,10 +129,10 @@ export class LisaServeIAMStack extends Stack {
     private createRagLambdaRole (config: Config): IRole {
         const ragLambdaRoleId = createCdkId([config.deploymentName, Roles.RAG_LAMBDA_EXECUTION_ROLE]);
         const ragLambdaRole = config.roles?.RagLambdaExecutionRole ?
-            Role.fromRoleName(this, ragLambdaRoleId, config.roles.RagLambdaExecutionRole) :
+            Role.fromRoleName(this.scope, ragLambdaRoleId, config.roles.RagLambdaExecutionRole) :
             this.createRagLambdaExecutionRole(config.deploymentName, ragLambdaRoleId);
 
-        new StringParameter(this, createCdkId(['LisaRagRole', 'StringParameter']), {
+        new StringParameter(this.scope, createCdkId(['LisaRagRole', 'StringParameter']), {
             parameterName: `${config.deploymentPrefix}/roles/${ragLambdaRoleId}`,
             stringValue: ragLambdaRole.roleArn,
             description: `Role ARN for LISA ${ragLambdaRoleId}`,
@@ -148,12 +143,12 @@ export class LisaServeIAMStack extends Stack {
 
     private createRagLambdaExecutionRole (deploymentName: string, roleName: string) {
         const lambdaPolicyStatements = getIamPolicyStatements('rag');
-        const lambdaRagPolicy = new ManagedPolicy(this, createCdkId([deploymentName, 'RAGPolicy']), {
+        const lambdaRagPolicy = new ManagedPolicy(this.scope, createCdkId([deploymentName, 'RAGPolicy']), {
             managedPolicyName: createCdkId([deploymentName, 'RAGPolicy']),
             statements: lambdaPolicyStatements,
         });
 
-        return new Role(this, Roles.RAG_LAMBDA_EXECUTION_ROLE, {
+        return new Role(this.scope, Roles.RAG_LAMBDA_EXECUTION_ROLE, {
             roleName,
             assumedBy: new ServicePrincipal('lambda.amazonaws.com'),
             description: 'Role used by RAG API lambdas to access AWS resources',
@@ -162,7 +157,7 @@ export class LisaServeIAMStack extends Stack {
     }
 
     private createEcsTaskRole (role: ECSRole, roleId: string, roleName: string, taskPolicy: IManagedPolicy): IRole {
-        return new Role(this, roleId, {
+        return new Role(this.scope, roleId, {
             assumedBy: new ServicePrincipal('ecs-tasks.amazonaws.com'),
             roleName,
             description: `Allow ${role.id} ${role.type} task access to AWS resources`,
