+++ lib/models/model-api.ts
@@ -46,7 +46,9 @@ import { UpdateModelStateMachine } from './state-machine/update-model';
 import { Secret } from 'aws-cdk-lib/aws-secretsmanager';
 import { createLambdaRole } from '../core/utils';
 import { Roles } from '../core/iam/roles';
+import * as path from 'path';
 
+const HERE = path.resolve(__dirname);
 /**
  * Properties for ModelsApi Construct.
  *
@@ -56,7 +58,7 @@ import { Roles } from '../core/iam/roles';
  * @property {ISecurityGroup[]} securityGroups - Security groups for Lambdas
  */
 type ModelsApiProps = BaseProps & {
-    authorizer: IAuthorizer;
+    authorizer?: IAuthorizer;
     lisaServeEndpointUrlPs: StringParameter;
     restApiId: string;
     rootResourceId: string;
@@ -188,12 +190,12 @@ export class ModelsApi extends Construct {
         };
 
         const lambdaRole: IRole = createLambdaRole(this, config.deploymentName, 'ModelApi', modelTable.tableArn, config.roles?.ModelApiRole);
+        const lambdaPath = path.join(HERE, '..', '..', 'lambda');
         // create proxy handler
         const lambdaFunction = registerAPIEndpoint(
             this,
             restApi,
-            authorizer,
-            './lambda',
+            lambdaPath,
             lambdaLayers,
             {
                 name: 'handler',
@@ -206,6 +208,7 @@ export class ModelsApi extends Construct {
             getDefaultRuntime(),
             vpc,
             securityGroups,
+            authorizer,
             lambdaRole,
         );
         lisaServeEndpointUrlPs.grantRead(lambdaFunction.role!);
@@ -271,13 +274,13 @@ export class ModelsApi extends Construct {
             registerAPIEndpoint(
                 this,
                 restApi,
-                authorizer,
-                './lambda',
+                lambdaPath,
                 lambdaLayers,
                 f,
                 getDefaultRuntime(),
                 vpc,
                 securityGroups,
+                authorizer,
                 lambdaRole,
             );
         });
