+++ lib/stages.ts
@@ -44,11 +44,10 @@ import { LisaServeApplicationStack } from './serve';
 import { UserInterfaceStack } from './user-interface';
 import { LisaDocsStack } from './docs';
 
-import path from 'node:path';
 import fs from 'node:fs';
+import { VERSION_PATH } from './util';
 
-const HERE: string = path.resolve(__dirname);
-const VERSION_PATH: string = path.resolve(HERE, '..', 'VERSION');
+
 export const VERSION: string = fs.readFileSync(VERSION_PATH, 'utf8').trim();
