+++ test/cdk/stacks/nag.test.ts
@@ -31,17 +31,17 @@ enum NagType {
 }
 
 const nagResults: NagResult = {
-    LisaApiBase: [1,7,0,5],
+    LisaApiBase: [1,7,0,7],
     LisaApiDeployment: [0,0,0,0],
-    LisaChat: [2,33,0,12],
+    LisaChat: [2,33,0,38],
     LisaCore: [0,0,0,0],
-    LisaDocs: [1,23,0,11],
+    LisaDocs: [1,23,0,13],
     LisaIAM: [0,14,0,0],
-    LisaModels: [1,74,0,22],
+    LisaModels: [1,74,0,62],
     LisaNetworking: [1,2,3,5],
-    LisaRAG: [2,66,0,21],
-    LisaServe: [1,21,0,29],
-    LisaUI: [0,16,0,6],
+    LisaRAG: [2,66,0,51],
+    LisaServe: [1,21,0,31],
+    LisaUI: [0,16,0,8],
 };
 
 describe('Nag Pack Tests', () => {
