+++ vector_store_deployer/.npmignore
@@ -1,0 +1,4 @@
+node_modules/
+src/
+package.json
+tsconfig.json
